move patched_httpx_client to conftest
