+++ tests/unit/conftest.py
@@ -7,6 +7,7 @@ from collections import defaultdict
 from logging import getLogger
 from typing import TYPE_CHECKING, Any, get_type_hints
 
+import httpx
 import pytest
 from pytest_httpserver import HTTPServer
 
@@ -209,3 +210,17 @@ def httpserver(make_httpserver: HTTPServer) -> Iterator[HTTPServer]:
     server = make_httpserver
     yield server
     server.clear()  # type: ignore[no-untyped-call]
+
+
+@pytest.fixture
+def patched_httpx_client(monkeypatch: pytest.MonkeyPatch) -> Iterator[None]:
+    """Patch httpx client to drop proxy settings."""
+
+    class ProxylessAsyncClient(httpx.AsyncClient):
+        def __init__(self, *args: Any, **kwargs: Any) -> None:
+            kwargs.pop('proxy', None)
+            super().__init__(*args, **kwargs)
+
+    monkeypatch.setattr(httpx, 'AsyncClient', ProxylessAsyncClient)
+    yield
+    monkeypatch.undo()
